Issue #2 - Restructured project to separate technology from intention
