--- src/main/g8/src/test/scala/uk/gov/hmrc/integration/cucumber/stepdefs/BaseStepDef.scala
@@ -1,21 +1,0 @@
-package uk.gov.hmrc.integration.cucumber.stepdefs
-
-import cucumber.api.scala.{EN, ScalaDsl}
-import uk.gov.hmrc.integration.cucumber.pages.BasePage._
-
-class BaseStepDef extends ScalaDsl with EN {
-
-  And("""^The user refreshes the page\$""") { () =>
-    pageRefresh()
-  }
-
-  And("""^the (Continue|Submit) button is clicked\$""") { () =>
-    clickContinue()
-  }
-
-  Then("""^the browser is shutdown\$""") { () =>
-    ShutdownTest()
-  }
-
-
-}
