--- src/main/g8/src/test/scala/uk/gov/hmrc/integration/cucumber/utils/Runner.scala
@@ -1,15 +1,0 @@
-package uk.gov.hmrc.integration.cucumber.utils
-
-import cucumber.api.CucumberOptions
-import cucumber.api.junit.Cucumber
-import org.junit.runner.RunWith
-
-@RunWith(classOf[Cucumber])
-@CucumberOptions(
-  features = Array("src/test/resources/features"),
-  glue = Array("uk.gov.hmrc.integration.cucumber.stepdefs"),
-  plugin = Array ("pretty", "html:target/cucumber", "json:target/cucumber.json"),
-  tags = Array("@Example, @Shutdown")
-)
-class Runner {
-}
