+++ src/main/g8/src/test/scala/uk/gov/hmrc/test/ui/conf/TestConfiguration.scala
@@ -1,10 +1,25 @@
-package uk.gov.hmrc.integration.cucumber.utils
+/*
+ * Copyright 2018 HM Revenue & Customs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package uk.gov.hmrc.test.ui.conf
 
 import java.io.FileInputStream
 import java.util.Properties
 
-object BaseUtil {
-  val fis = new FileInputStream("src/test/scala/uk/gov/hmrc/integration/cucumber/environments.properties")
+object TestConfiguration {
+  val fis = new FileInputStream("src/test/resources/environments.properties")
   val props: Properties = new Properties()
   loadProperties(fis, props)
