+++ src/main/g8/src/test/scala/uk/gov/hmrc/test/ui/cucumber/stepdefs/ExampleStepDef.scala
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016 HM Revenue & Customs
+ * Copyright 2018 HM Revenue & Customs
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,11 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
-package uk.gov.hmrc.integration.cucumber.stepdefs
+package uk.gov.hmrc.test.ui.cucumber.stepdefs
 
 import cucumber.api.scala.{EN, ScalaDsl}
-import uk.gov.hmrc.integration.cucumber.pages.ExamplePage._
+import uk.gov.hmrc.test.ui.pages.ExamplePage._
 
 class ExampleStepDef extends ScalaDsl with EN {
