+++ src/main/g8/src/test/scala/uk/gov/hmrc/test/ui/runner/driver/browsers/ChromeBrowser.scala
@@ -1,9 +1,24 @@
-package uk.gov.hmrc.integration.cucumber.utils.driver.browsers
+/*
+ * Copyright 2018 HM Revenue & Customs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package uk.gov.hmrc.test.ui.runner.driver.browsers
 
 import org.openqa.selenium.WebDriver
 import org.openqa.selenium.chrome.{ChromeDriver, ChromeOptions}
 import org.openqa.selenium.remote.{CapabilityType, DesiredCapabilities}
-import uk.gov.hmrc.integration.cucumber.utils.driver.ProxySupport
+import uk.gov.hmrc.test.ui.runner.driver.ProxySupport
 
 object ChromeBrowser extends ProxySupport {
